+++ modules/tinymce/src/core/main/ts/fmt/Preview.ts
@@ -285,7 +285,7 @@ const getCssText = (editor: Editor, format: any) => {
 
       // Ignore white since it's the default color, not the nicest fix
       // TODO: Fix this by detecting runtime style
-      if (Transformations.anyToHexString(value).toLowerCase() === '#ffffff') {
+      if (Transformations.rgbaToHexString(value).toLowerCase() === '#ffffff') {
         return;
       }
     }
@@ -293,7 +293,7 @@ const getCssText = (editor: Editor, format: any) => {
     if (name === 'color') {
       // Ignore black since it's the default color, not the nicest fix
       // TODO: Fix this by detecting runtime style
-      if (Transformations.anyToHexString(value).toLowerCase() === '#000000') {
+      if (Transformations.rgbaToHexString(value).toLowerCase() === '#000000') {
         return;
       }
     }